+++ db/db_sst_test.cc
@@ -701,9 +701,13 @@ TEST_F(DBSSTTest, GetTotalSstFilesSize) {
   ASSERT_TRUE(dbfull()->GetIntProperty("rocksdb.total-sst-files-size",
                                        &total_sst_files_size));
   // Live SST files = 1 (compacted file)
-  // Total SST files = 6 (5 original files + compacted file)
-  ASSERT_EQ(live_sst_files_size, 1 * single_file_size);
-  ASSERT_EQ(total_sst_files_size, 6 * single_file_size);
+  // The 5 bytes difference comes from oldest-key-time table property isn't
+  // propagated on compaction. It is written with default value
+  // std::numeric_limits<uint64_t>::max as varint64.
+  ASSERT_EQ(live_sst_files_size, 1 * single_file_size + 5);
+
+  // Total SST files = 5 original files + compacted file
+  ASSERT_EQ(total_sst_files_size, 5 * single_file_size + live_sst_files_size);
 
   // hold current version
   std::unique_ptr<Iterator> iter2(dbfull()->NewIterator(ReadOptions()));
@@ -724,14 +728,14 @@ TEST_F(DBSSTTest, GetTotalSstFilesSize) {
                                        &total_sst_files_size));
   // Live SST files = 0
   // Total SST files = 6 (5 original files + compacted file)
-  ASSERT_EQ(total_sst_files_size, 6 * single_file_size);
+  ASSERT_EQ(total_sst_files_size, 5 * single_file_size + live_sst_files_size);
 
   iter1.reset();
   ASSERT_TRUE(dbfull()->GetIntProperty("rocksdb.total-sst-files-size",
                                        &total_sst_files_size));
   // Live SST files = 0
   // Total SST files = 1 (compacted file)
-  ASSERT_EQ(total_sst_files_size, 1 * single_file_size);
+  ASSERT_EQ(total_sst_files_size, live_sst_files_size);
 
   iter2.reset();
   ASSERT_TRUE(dbfull()->GetIntProperty("rocksdb.total-sst-files-size",