+++ include/rocksdb/table_properties.h
@@ -4,8 +4,9 @@
 #pragma once
 
 #include <stdint.h>
-#include <string>
+#include <limits>
 #include <map>
+#include <string>
 #include "rocksdb/status.h"
 #include "rocksdb/types.h"
 
@@ -49,6 +50,7 @@ struct TablePropertiesNames {
   static const std::string kPropertyCollectors;
   static const std::string kCompression;
   static const std::string kCreationTime;
+  static const std::string kOldestKeyTime;
 };
 
 extern const std::string kPropertiesBlock;
@@ -162,6 +164,8 @@ struct TableProperties {
   // The time when the SST file was created.
   // Since SST files are immutable, this is equivalent to last modified time.
   uint64_t creation_time = 0;
+  // Timestamp of the earliest key
+  uint64_t oldest_key_time = std::numeric_limits<uint64_t>::max();
 
   // Name of the column family with which this SST file is associated.
   // If column family is unknown, `column_family_name` will be an empty string.